feat: LCD 기능 추가 - LCD에 온도 및 기타 상태 업데이트 기능 추가 - TODO  + 온도 설정 변경 기능 추가

+++ lib/libtc/include/ce/tc/tc_state.h
@@ -10,7 +10,7 @@ typedef struct {
     uint8_t wifi_connected;
     uint8_t alarm_connected;
     uint8_t warning_connected;
-    uint8_t machine_condition;
+    float machine_condition;
     SemaphoreHandle_t tc_state_mutex;
 } tc_state_t;
